(WIP) Moved common interfaces under view package.
