+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/views/Presenter.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.commons;
+package org.hisp.dhis.android.dataentry.commons.views;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.UiThread;
