+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/views/View.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.commons;
+package org.hisp.dhis.android.dataentry.commons.views;
 
 public interface View {
 }
